Install new icon set to follow GNOME standards
The pybliographer icon had to be redrawn and replaced, because
there was no svg version available. (It's a quick solution, the
result might not be ideal :). The icons in different sizes are
installed now in hicolor icon dir.
The pixmaps used by the program internally are moved under the
package datadir for the moment.